+++ backend/src/database/migrations/20220408185000_create_subcontractors_table.js
@@ -1,0 +1,10 @@
+exports.up = function (knex) {
+  return knex.schema.createTable("subcontractors", function (table) {
+    table.increments(); //id
+    table.string("name").notNullable();
+  });
+};
+
+exports.down = function (knex) {
+  return knex.schema.dropTable("subcontractors");
+};